requirements.txt
python-dotenv==1.0.0
webdriver_manager==3.8.6
playwright==1.48.0
APScheduler==3.10.4
Flask==3.0.0
psutil==5.9.5
